Make an example test and service to test to show the reason for onMockFor
This commit contains a test file with a simplified example of a service one wants to test and shows how onMockFor is used to test it.

I was not able to use existing Mocker functionality to create the test to test this functionality.  Might be a way, but I’m not seeing it. Hence onMockFor which, as you can see in this test, works perfectly.